Authorizations for orders, small change in security.yaml for the paths that anyone can access

+++ src/Controller/OrderController.php
@@ -10,6 +10,7 @@ use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpFoundation\Response;
 use Symfony\Component\Routing\Attribute\Route;
+use function Symfony\Component\Clock\now;
 
 final class OrderController extends AbstractController
 {
@@ -37,6 +38,9 @@ final class OrderController extends AbstractController
     {
         $order = new Order();
 
+        $order->setConsumer($this->getUser()->getConsumer());
+        $order->setCreated_at(now());
+
         $form = $this->createForm(OrderFormType::class, $order);
         $form->handleRequest($request);
 
@@ -55,6 +59,10 @@ final class OrderController extends AbstractController
     #[Route('/order/edit/{id}', name: 'app_order_edit', methods: ['GET', 'POST'])]
     public function edit(Request $request, Order $order, EntityManagerInterface $entityManager): Response
     {
+        if ($this->getUser()->getConsumer() !== $order->getConsumer()) {
+            return $this->redirectToRoute('app_order');
+        }
+
         $form = $this->createForm(OrderFormType::class, $order);
         $form->handleRequest($request);
 
@@ -69,10 +77,18 @@ final class OrderController extends AbstractController
         ]);
     }
 
-    #[Route('/order/delete/{id}', name: 'app_order_delete', methods: ['GET','POST'])]
+    #[Route('/order/delete/{id}', name: 'app_order_delete', methods: ['GET', 'POST'])]
     public function delete(Request $request, int $id, OrderRepository $orderRepository, EntityManagerInterface $entityManager): Response
     {
+        $thisConsumer = $this->getUser()->getConsumer();
+        $thisBusiness = $this->getUser()->getBusiness();
         $order = $orderRepository->find($id);
+        if (($thisConsumer !== null && $thisConsumer !== $order->getConsumer()) ||
+            ($thisBusiness !== null && $thisBusiness !== $order->getBusiness())) {
+            return $this->redirectToRoute('app_order');
+        }
+
+
         $entityManager->remove($order);
         $entityManager->flush();
